+++ satella/coding/sequences/__init__.py
@@ -1,6 +1,6 @@
 from .choose import choose
-from .iterators import infinite_counter, take_n, is_instance, skip_first, zip_shifted
+from .iterators import infinite_counter, take_n, is_instance, skip_first, zip_shifted, stop_after
 from .sequences import is_last, add_next, half_product
 
 __all__ = ['choose', 'infinite_counter', 'take_n', 'is_instance', 'is_last', 'add_next',
-           'half_product', 'skip_first', 'zip_shifted']
+           'half_product', 'skip_first', 'zip_shifted', 'stop_after']